Update README.md
Choose an appropriate AWS service: AWS offers multiple services for hosting static websites, such as Amazon S3.  Now we can choose the one that best fits our requirements.

Create an S3 bucket: We can create an S3 bucket and configure it to host our website. We should ensure that the S3 bucket is not publicly accessible, and enable website hosting.

Upload website content to the S3 bucket: We can upload our website content to the S3 bucket, such as HTML, CSS, and JavaScript files.

Use CloudFront for content delivery: We can use AWS CloudFront to distribute our website content to users worldwide, which also provides added security benefits. CloudFront can also be configured to redirect HTTP requests to HTTPS.

Use AWS Certificate Manager to generate SSL certificate: We can use AWS Certificate Manager to create a free SSL certificate for our domain.

Configure a web server: We can use a configuration management tool such as Ansible or Terraform to deploy a web server such as Apache or Nginx, and configure it to serve the static website content.

Secure the web server: We can configure the webserver to redirect HTTP requests to HTTPS, and limit access to the appropriate ports.

Apply automated tests: We can use a testing tool such as Selenium or Puppeteer to develop and run automated tests to ensure that our server configuration is correct.

Express everything in code: To automate the deployment and configuration process, we can express everything in code using infrastructure as code tools such as AWS CloudFormation or Terraform.